test_requirements.txt
mock
pytest
pytest-cov
pytest-django
requests
